Criacao de componente para gerencia de todos os usuario pelos ADMs

+++ src/app/home/home.component.ts
@@ -5,6 +5,8 @@ import { CategoriaService } from './../service/categoria.service';
 import { Produto } from './../model/Produto';
 import { ProdutoService } from './../service/produto.service';
 import { Component, OnInit } from '@angular/core';
+import { environment } from 'src/environments/environment.prod';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
   selector: 'app-home',
@@ -14,14 +16,19 @@ import { Component, OnInit } from '@angular/core';
 export class HomeComponent implements OnInit {
 
   public listaCategorias: Categoria[];
+  public listaMesas: Mesa[];
 
   public mesa: Mesa = new Mesa();
   public produto: Produto = new Produto();
 
+  public id: number = environment.id;
+  public tipo: string = environment.tipo;
+
   constructor(
     private produtoService: ProdutoService,
     private categoriaService: CategoriaService,
-    private mesaService: MesaService
+    private mesaService: MesaService,
+    private route: ActivatedRoute
 
   ) { }
 
@@ -30,7 +37,16 @@ export class HomeComponent implements OnInit {
 
     this.getAllCategorias(); // CARREGA A LISTA COM TODOS OS PRODUTOS DA BASE DE DADOS
 
-    this.getByIdMesa(2); // VALOR DE ID FORCADO PARA TESTE
+    if(this.tipo != "mesa") {
+      // IRA ENTRAR NESSA CONDICAO CASO SEJA UM ADM
+      this.id = 2//this.route.snapshot.params['id'];
+
+    }else {
+      this.id = environment.id;
+
+    }
+
+    this.getByIdMesa(this.id); // VALOR DE ID FORCADO PARA TESTE
 
   }
 
@@ -43,7 +59,7 @@ export class HomeComponent implements OnInit {
   }
 
   getByIdMesa(id: number) {
-    this.mesaService.findByIdCategoria(id).subscribe((resp: Mesa) => {
+    this.mesaService.findByIdMesa(id).subscribe((resp: Mesa) => {
 
       // NAVEGA NO RETORNO DA MESA, E PREENCHE O NOME DO PRODUTO
       resp.produtos.map((item) => {
@@ -90,20 +106,27 @@ export class HomeComponent implements OnInit {
   }
 
   removerProduto(idMesa: number, idProduto: number) {
-    this.mesa.produtos = []; // ZERA A MEMORIA DO OBJ PARA QUE POSSA SER POPULADO NOVAMENTE COM OS VALORES DO BANCO
-    this.mesa.total = 0; // ZERA A MEMORIA DO OBJ PARA QUE POSSA SER POPULADO NOVAMENTE COM OS VALORES DO BANCO
+    if(this.tipo != "mesa") {
 
-    this.mesaService.removerProdutoAMesa(idMesa, idProduto).subscribe((resp: boolean) => {
+      this.mesa.produtos = []; // ZERA A MEMORIA DO OBJ PARA QUE POSSA SER POPULADO NOVAMENTE COM OS VALORES DO BANCO
+      this.mesa.total = 0; // ZERA A MEMORIA DO OBJ PARA QUE POSSA SER POPULADO NOVAMENTE COM OS VALORES DO BANCO
 
-      if(resp) {
-        this.getByIdMesa(idMesa);
+      this.mesaService.removerProdutoAMesa(idMesa, idProduto).subscribe((resp: boolean) => {
 
-      }else {
-        alert("Ocorreu um erro ao tentar remover o produto no seu carrinho.");
+        if(resp) {
+          this.getByIdMesa(idMesa);
 
-      }
+        }else {
+          alert("Ocorreu um erro ao tentar remover o produto no seu carrinho.");
 
-    });
+        }
+
+      });
+
+    }else {
+      alert("Você não tem permissão para realizar essa ação.");
+
+    }
 
   }
